Reorganize ActionCreator tests and cover implicit actions (#28332)
* Remove action creator's `setupEditing` test util

* Prepare to break down action creator tests

* Add tests for implicit actions

* Extract common tests

* Extract sharing tests

* Move utils tests to tests dir

* Fix tests

* Use `createAction` E2E helper

* Extract public sharing E2E helpers

* Add E2E test for public implicit action
